+++ presto-spi/src/main/java/com/facebook/presto/spi/type/LongDecimalType.java
@@ -106,7 +106,7 @@ public final class LongDecimalType
     }
 
     @Override
-    public int hash(Block block, int position)
+    public long hash(Block block, int position)
     {
         return block.hash(position, 0, LONG_DECIMAL_LENGTH);
     }
